+++ config/test.exs
@@ -12,6 +12,7 @@ config :tesla, Lightning.VersionControl.GithubClient,
 config :tesla, Mix.Tasks.Lightning.InstallAdaptorIcons, adapter: Tesla.Mock
 
 config :tesla, Lightning.UsageTracking.Client, adapter: Tesla.Mock
+config :tesla, Lightning.UsageTracking.GithubClient, adapter: Tesla.Mock
 
 # Configure your database
 #