+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to
 
 - Change all System.get_env calls in runtime.exs to use dotenvy
   [#1968](https://github.com/OpenFn/lightning/issues/1968)
+- Send richer version info as part of usage tracking submission.
+  [#1819](https://github.com/OpenFn/lightning/issues/1819)
 
 ### Fixed